test: 테스트 패키지 분리하도록 수정
